% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_plot.R
\name{activity_plot}
\alias{activity_plot}
\title{Title}
\usage{
activity_plot(
  input.path,
  scream_thres,
  filter_date,
  start_date = 0,
  end_date = 0
)
}
\arguments{
\item{input.path}{Path to input data}

\item{scream_thres}{Scream threshold (a value higher than the threshold will alarm the user)}

\item{filter_date}{If filter date is activated only data within a user defined period is analysed}

\item{start_date}{Start of the user defined time period of interest}

\item{end_date}{End of the user defined time period of interest}
}
\value{
Create activity plot of detected pig screams
}
\description{
Title
}
\examples{
activity_plot(out.path,scream_thres,filter_date,start_date,end_date)
}
