% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_seconds_to_d_h_m_s_ms.R
\name{convert_seconds_to_d_h_m_s_ms}
\alias{convert_seconds_to_d_h_m_s_ms}
\title{Converts seconds into days/hours/minutes/seconds/milliseconds}
\usage{
convert_seconds_to_d_h_m_s_ms(x)
}
\arguments{
\item{x}{time value (seconds) that must be converted}
}
\value{
Return converted time format
}
\description{
Converts seconds into days/hours/minutes/seconds/milliseconds
}
\examples{
convert_seconds_to_d_h_m_s_ms(x)
}
