% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logbook.R
\name{create_logbook}
\alias{create_logbook}
\title{Create log book entry}
\usage{
create_logbook(out.path, log_entry)
}
\arguments{
\item{out.path}{Output directory path}

\item{log_entry}{Logbook data}
}
\value{
Return log book entry
}
\description{
Create log book entry
}
\examples{
create_logbook(out.path,log_entry)
}
