% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mp4_of_calls.R
\name{create_mp4_of_calls}
\alias{create_mp4_of_calls}
\title{Create video files of detected calls}
\usage{
create_mp4_of_calls(
  pig_scream_table,
  audio_ref_filter,
  mp4_path,
  mp4_name,
  mp4_dur,
  out.dir,
  wav_dur
)
}
\arguments{
\item{pig_scream_table}{Pig Scream Table data}

\item{audio_ref_filter}{Reference filter status (on/off)}

\item{mp4_path}{Path to mp4 data}

\item{mp4_name}{Names of mp4 data}

\item{mp4_dur}{Duration of mp4 calls}

\item{out.dir}{Directory of output folder}

\item{wav_dur}{Audio file duration}
}
\value{
Return video files of detected calls
}
\description{
Create video files of detected calls
}
\examples{
create_mp4_of_calls(pig_scream_table,audio_ref_filter,mp4_path,mp4_name,mp4_dur,out.dir,wav_dur,cor_temps=NULL)
}
