% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase_data.R
\name{erase_data}
\alias{erase_data}
\title{Title}
\usage{
erase_data(input_path, erase_date)
}
\arguments{
\item{input_path}{Directory to input data}

\item{erase_date}{Erasing date (all output data of the detector from recordings older than the erasing date is erased)}
}
\value{
Erase old data (user defined)
}
\description{
Title
}
\examples{
erase_data(out.path,erase_date)
}
