% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_new_mp4.R
\name{find_new_mp4}
\alias{find_new_mp4}
\title{Filter mp4 for files created in the last hour}
\usage{
find_new_mp4(mp4_files, time_window, get_all_mp4)
}
\arguments{
\item{mp4_files}{List of mp4 files}

\item{time_window}{Time window for definition new data}

\item{get_all_mp4}{Switch to decide if all or only a specific set mp4 is analyzed}
}
\value{
List of newly created mp4
}
\description{
Filter mp4 for files created in the last hour
}
\examples{
find_new_mp4(mp4_files,time_window)
}
