% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_detections.R
\name{get_detections}
\alias{get_detections}
\title{Searches for matches between templates and detected calls}
\usage{
get_detections(wav_path, cor_temps)
}
\arguments{
\item{wav_path}{Path to wave data}

\item{cor_temps}{Template data}
}
\value{
Return matches between calls and templates
}
\description{
Searches for matches between templates and detected calls
}
\examples{
get_detections(wav_path,cor_temps)
}
