% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mp4_files.R
\name{get_mp4_files}
\alias{get_mp4_files}
\title{Get all mp4 files in folder "surveillance"}
\usage{
get_mp4_files(work_path, keywords)
}
\arguments{
\item{work_path}{A path to the input mp4 files}

\item{keywords}{Add search keywords to identify mp4 directories}
}
\value{
List of mp4 file locations
}
\description{
Get all mp4 files in folder "surveillance"
}
\examples{
get_mp4_files(work_path)
}
