% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_pig_screams.R
\name{identify_pig_screams}
\alias{identify_pig_screams}
\title{Identify pig screams that fit the parameters provided by the user}
\usage{
identify_pig_screams(
  wav_path,
  filter,
  lower_cut_off_freq,
  higher_cut_off_freq,
  loudness_threshold,
  time_gap,
  call_dur,
  ref_switch,
  scream_threshold = NULL
)
}
\arguments{
\item{wav_path}{Path to wave files}

\item{filter}{Information if audio filter function is active}

\item{lower_cut_off_freq}{Lower cutoff frequency}

\item{higher_cut_off_freq}{Higher cutoff frequency}

\item{loudness_threshold}{Loudness threshold (only calls above the threshold value are detected)}

\item{time_gap}{Time gap between calls}

\item{call_dur}{Minimal duration time a call must have}

\item{ref_switch}{Reference switch (if on reference audio files are created)}

\item{scream_threshold}{Scream threshold (a warning message is send if the number of detected calls is above the threshold)}
}
\value{
Return table with pig screams
}
\description{
Identify pig screams that fit the parameters provided by the user
}
\examples{
identify_pig_screams(wav_path,filter,lower_cut_off_freq,higher_cut_off_freq,loudness_threshold,time_gap,call_dur,ref_switch,scream_threshold=NULL)
}
