% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ref_thres.R
\name{identify_ref_thres}
\alias{identify_ref_thres}
\title{Calculate call detection threshold based on reference calls}
\usage{
identify_ref_thres(
  work_path,
  ref.files,
  ref_call,
  filter,
  lower_cut_off_freq,
  higher_cut_off_freq,
  loudness_threshold
)
}
\arguments{
\item{work_path}{Path to data}

\item{ref.files}{Name list of reference files}

\item{ref_call}{Name list of reference calls}

\item{filter}{Filter}

\item{lower_cut_off_freq}{Lower cutoff frequency parameter}

\item{higher_cut_off_freq}{Higher cutoff frequency parameter}

\item{loudness_threshold}{Loudness threshold}
}
\value{
Detected pig scream calls
}
\description{
Calculate call detection threshold based on reference calls
}
\examples{
identify_ref_thres(work_path,ref.files,ref_call,filter,lower_cut_off_freq,higher_cut_off_freq,loudness_threshold)
}
