% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_program_parameters.R
\name{load_program_parameters}
\alias{load_program_parameters}
\title{Load parameters from folder "parameters"}
\usage{
load_program_parameters(path)
}
\arguments{
\item{path}{Path to project folder}
}
\value{
Return user parameters
}
\description{
Load parameters from folder "parameters"
}
\examples{
load_program_parameters()


}
