% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp4_to_wav_converter.R
\name{mp4_to_wav_converter}
\alias{mp4_to_wav_converter}
\title{Convert mp4 to wav format}
\usage{
mp4_to_wav_converter(mp4_paths)
}
\arguments{
\item{mp4_paths}{Path to mp4 file}
}
\value{
wav files (from mp4)
}
\description{
Convert mp4 to wav format
}
\examples{
mp4_to_wav_converter(mp4_paths)
}
