% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreamDetect.R
\name{ScreamDetect}
\alias{ScreamDetect}
\title{Title ScreamDetect}
\usage{
ScreamDetect(path)
}
\arguments{
\item{path}{Path to R script}
}
\value{
Detected pig screams
}
\description{
Title ScreamDetect
}
\examples{
ScreamDetect("path_to_R_script")
ScreamDetect("C:/Users/tjard/OneDrive/Desktop/PSD/")
}
