% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_message.R
\name{send_message}
\alias{send_message}
\title{Send warning email}
\usage{
send_message(table.path, data_name, pig_scream_table, mp4_name, email_receiver)
}
\arguments{
\item{table.path}{Path to table}

\item{data_name}{Name of data}

\item{pig_scream_table}{Table that lists detected calls}

\item{mp4_name}{Name of mp4 file}

\item{email_receiver}{email address or adresses of data recipients}
}
\value{
Return warning email
}
\description{
Send warning email
}
\examples{
send_message(table.path,data_name,pig_scream_table,mp4_name,email_receiver)
}
