% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_scheduler_screamdetect.R
\name{task_scheduler_psd}
\alias{task_scheduler_psd}
\title{Title Task Scheduler for Pig Scream Detector}
\usage{
task_scheduler_psd(path, script_name)
}
\arguments{
\item{path}{Path to detection project folder}

\item{script_name}{Name of R script that will be executed on a user defined time schedule}
}
\value{
Set time schedule for automated pig scream detection
}
\description{
Title Task Scheduler for Pig Scream Detector
}
\examples{
task_scheduler_psd("C:/Users/tjard/OneDrive/Desktop/PSD_Project/","psd.R")
task_scheduler_psd(path,script_name)
}
